requirements.txt
blinker==1.9.0
click==8.1.7
cloudevents==1.11.0
colorama==0.4.6
deprecation==2.1.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
parliament-functions==0.1.0
waitress==3.0.2
Werkzeug==3.1.3